backend/requirements.txt
Django==5.0.6
django-rest-framework==0.1.0
python-dotenv==1.0.1
requests==2.32.3
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
pymemcache==4.0.0
gunicorn==23.0.0
openai==1.52.2
drf-nested-routers==0.94.1
psycopg==3.2.4
ruff==0.9.6
drf-spectacular==0.28.0
Pillow
weasyprint==64.1
celery==5.5.3
redis>=4.0
django-celery-beat>=2.0
flower
cryptography
stripe==10.12.0
twilio==8.1.0
